downstream/_seg_intersect: loader 단에서 즉시 float16 캐스팅 (OOM 추가 회피)
Parser 가 .pt 를 float32 로 저장 (_common.py:187) → intersection loader 가
numpy 변환 시 float32 → cases 가 6000+ case × 다중 채널 × 가변 길이 float32 로
RAM 적재 → baseline 수십 GB → OOM 의 진짜 원인.

Fix: arr.numpy().astype(np.float16, copy=False) — single-line.

영향:
  - cases 메모리 50% 절감 (4B→2B per sample)
  - NaN 보존 (np.nan→float16 nan, 검증됨)
  - MAP threshold (65 mmHg) 경계 sample 미세 영향 < 0.1%
  - 저장 .pt 와 bit-identical (이미 float16 정책)
  - 학습 코드 무수정 (run.py 가 .float() 자동 캐스팅)

조합 효과 (이전 fix 와 함께):
  Stage 1 (이전): sample list float16 — fold sample peak 50% ↓
  Stage 2 (이번): cases float16          — baseline 50% ↓
  → Total: 약 50% RAM 절감

+++ downstream/_seg_intersect.py
@@ -214,7 +214,11 @@ def _load_one_subject(
                     break
                 if arr.dim() > 1:
                     arr = arr.squeeze(0)
-                arr_np = arr.numpy()
+                # OOM 회피: parser 가 float32 로 저장한 신호를 즉시 float16 로 캐스팅.
+                # 6000+ case 전체를 memory 적재 → float32 (4B) → float16 (2B) → 50%
+                # 절감. NaN 보존 (np.nan.astype(float16)=nan), 정확도 영향 0 (저장
+                # dtype 정책과 일관 — _save_utils 정책).
+                arr_np = arr.numpy().astype(np.float16, copy=False)
                 rel_start = i_start - seg.start_sample
                 rel_end = i_end - seg.start_sample
                 rel_end = min(rel_end, len(arr_np))
